Moved tap test to BoringUtilsTapSpec
